Add the database server
Use the default postgres-13 image. Initialize the database using 'koji'
as the database name and 'koji' as the admin user.

This version of postgres is behind the version of postgres shipped in
Fedora 36 due to an incompatibility between postgres 14 and koji-1.28.

Add a script to generate shared secrets. Create a random password for
the koji database to be shared between the database server and koji-hub.