Give up on the koji-web -> koji-hub reverse proxy
Since the koji-web to koji-hub connection also needs SSL verification,
there's not a way to pass the client certificate (of the user using the
CLI to access koji-hub) through to koji-hub in a way that doesn't create
a gigantic mess. Just expose koji-hub on a different port and use that
instead.

Redo the koji-hub certificate to use subject alternative names to make
this work as expected with the koji client.