Fix some issues around binfmt_misc and setuid
Add instructions for how to setup binfmt_misc for running non-readable
setuid binaries, which is needed by the consolehelper/userhelper
wrappers used by mock.

Add a patch to remove a setregid call that causes qemu to fail when mock
runs it from inside a child process in `update_config_from_file`. I
don't entirely understand how this is causing things to fail, I just
know it is, and I also know how to remove it.