Add robosignatory
Modify the README to use a separate destination tag to place unsigned
builds.

Connect robosignatory to the sigul-bridge.

Set robosignatory as the sigul-server admin. Creating separate users for
admin and key owner would require adding non-admin users to sigul, and
there's not a tool for that, and also it doesn't matter that much.

Modify the sigul-bridge koji config to match the koji instances expected
by robosignatory.