Remove the koji-admin container
Move the koji user and host bootstrap to the koji-admin entrypoint.
Handle the chicken-and-egg problem of can't run the koji CLI without
koji-hub by just running the SQL manually. Move the koji-admin
certificate to the filesystem and modify the README to run the
koji-admin commands from the host. Remove that awful probe thing in the
builders.